[ChartEditor] 차트제목 표기
